[cli] Initialize solve_poly()

+++ cli/main.h
@@ -25,7 +25,7 @@ void MakePoly(pList);           // Make a new polynomial
 void ShowList(pList, int);           // Display all polynomials
 pPoly AddPoly(pPoly, pPoly);       // Add two polynomials
 pPoly SubPoly(pPoly, pPoly);       // Sub two polynomials
-int SolvePoly(pList);      // Solve the polynomial
+double SolvePoly(pList);      // Solve the polynomial
 void DelPoly(pPoly);            // Delete the whole polynomial
 poly EmptyPoly(poly);           // Make one polynomial empty
 void ChangePoly(pList);          // Make change to one node of one polynomial
